raspi数字人1215/requirements.txt
opencv-python
pillow
pygame
pyaudio
numpy
requests
dashscope
baidu-aip
